+++ ssaHAL/ssa/__init__.py
@@ -1,4 +1,5 @@
 from .core import SSA
 from .decorators import ssa_task, ssa_main
+from .interfaces import NetworkDriver
 
-__all__ = ["SSA", "ssa_task", "ssa_main"]
+__all__ = ["SSA", "ssa_task", "ssa_main", "NetworkDriver"]
