+++ ssaHAL/ssa_modules/__init__.py
@@ -1,0 +1,3 @@
+from .config import ConfigLoader
+
+__all__ = ['ConfigLoader']
